- some more permissions

+++ app/views/submission/index.blade.php
@@ -21,7 +21,7 @@
 			<td style="width: 25%;"><strong>Option</strong></td>
 		</tr> 
 		@foreach ($submissions as $sub) 
-			<tr>
+			<tr class="@if ($sub->status === 1) success @elseif ($sub->status === 9) danger @endif">
 				<td>{{ link_to_route('submission.show', $sub->sub_title, [$sub->sub_id], null)}}</td>
 				<td>
 					@if ($sub->sub_type === 3)
@@ -36,19 +36,21 @@
 				<td>{{ date("d F Y",strtotime($sub->created_at)) }} at {{ date("g:ha",strtotime($sub->created_at)) }}</td>
 				<td>
 					@if ($sub->status === 1)
-					    <span class="text-success">Accepted</span>
+					    <span class="text-success"><strong>Accepted</strong></span>
 					@elseif ($sub->status === 9)
-					    <span class="text-danger">Rejected</span>
+					    <span class="text-danger"><strong>Rejected</strong></span>
 					@else
 					    On review
 					@endif
 				</td>
 				<td>
 					{{ link_to_route('submission.reviews', 'Reviews', [$sub->sub_id], ['class' => 'btn btn-default btn-xs'])}}
-					{{ link_to_route('submission.show', 'Edit Submission', [$sub->sub_id], ['class' => 'btn btn-info btn-xs'])}}
-					{{ Form::model($sub, ['route' => ['submission.destroy', $sub->sub_id], 'method' => 'delete', 'class' => 'inline' ]) }}
-						{{ Form::button('Withdraw', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs'])}}
-					{{ Form::close() }}
+					{{ link_to_route('submission.show', 'View/Edit Submission', [$sub->sub_id], ['class' => 'btn btn-info btn-xs'])}}
+					@if ($sub->status === 0)
+						{{ Form::model($sub, ['route' => ['submission.destroy', $sub->sub_id], 'method' => 'delete', 'class' => 'inline' ]) }}
+							{{ Form::button('Withdraw', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs'])}}
+						{{ Form::close() }}
+					@endif
 				</td>
 			</tr>
 		@endforeach
